requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster==0.7.13
anndata==0.9.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.2.4
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
black==24.8.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
cell2sentence==1.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cmake==3.31.4
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
datasets==3.1.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@b1c43ffbc995426a9a6b5c667730091a384e0fa4
dill==0.3.8
docker-pycreds==0.4.0
docutils==0.20.1
dominate==2.9.1
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.13
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
flake8==7.1.1
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gensim==4.3.3
get-annotations==0.1.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.69.0
h11==0.14.0
h5py==3.11.0
HTML4Vision==0.5.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
igraph==0.11.8
imageio==2.35.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.12.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
joblib==1.4.2
json5==0.11.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lightning-utilities==0.11.9
lit==18.1.8
llvmlite==0.41.1
lyft-dataset-sdk==0.0.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
mizani==0.9.3
mmcv-full==1.4.0
mmdet==2.14.0
mmdet3d==0.17.1
mmengine==0.10.6
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nlg-eval==2.3
nltk==3.9.1
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
plotnine==0.12.4
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.2.7
pynndescent==0.5.13
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.4
python-dateutil==2.9.0.post0
python-igraph==0.11.8
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.4.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.20.1
rsa==4.9
ruff==0.8.6
sacremoses==0.1.1
safetensors==0.5.3
scanpy==1.9.8
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.25.1
session-info==1.0.0
setproctitle==1.3.5
Shapely==1.8.5.post1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.1.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
texttable==1.7.0
Theano==1.0.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.12
tinycss2==1.2.1
tokenizers==0.12.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.4.1
torchaudio==0.9.1
torchmetrics==1.5.2
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
triton==3.0.0
typer==0.15.1
typing_extensions==4.13.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.2.3
uvicorn==0.33.0
wandb==0.19.9
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wrapt==1.17.0
xdg==6.0.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.1
yarl==1.15.2
zipp==3.20.2
